# update kurl_proxy to use stronger cipher suites